% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs_auxiliares.R
\encoding{UTF-8}
\name{Tabelas_Auxiliares}
\alias{Tabelas_Auxiliares}
\alias{consultar_tabela_auxiliar}
\alias{pesquisar_ncm}
\alias{pesquisar_sh6}
\alias{pesquisar_sh4}
\alias{pesquisar_sh2}
\alias{pesquisar_ncm_sec}
\alias{pesquisar_pais}
\alias{pesquisar_blocos}
\alias{pesquisar_uf}
\alias{pesquisar_mun}
\alias{pesquisar_via}
\alias{pesquisar_urf}
\alias{pesquisar_cgce3}
\alias{pesquisar_cgce2}
\alias{pesquisar_cgce1}
\alias{pesquisar_cuci_item}
\alias{pesquisar_cuci_sub}
\alias{pesquisar_cuci_pos}
\alias{pesquisar_cuci_cap}
\alias{pesquisar_cuci_sec}
\alias{pesquisar_isic_classe}
\alias{pesquisar_isic_grupo}
\alias{pesquisar_isic_div}
\alias{pesquisar_isic_sec}
\title{Tabelas auxiliares do Comex Stat}
\usage{
consultar_tabela_auxiliar(
  pattern = "",
  endpoint = stop("endpoint obrigatório"),
  filter_id = stop("filter_id obrigatório"),
  col_names = c("ID", "Descrição"),
  trim_chars = NULL,
  distinct_id = FALSE
)

pesquisar_ncm(pattern = "")

pesquisar_sh6(pattern = "")

pesquisar_sh4(pattern = "")

pesquisar_sh2(pattern = "")

pesquisar_ncm_sec(pattern = "")

pesquisar_pais(pattern = "")

pesquisar_blocos(pattern = "")

pesquisar_uf(pattern = "")

pesquisar_mun(pattern = "")

pesquisar_via(pattern = "")

pesquisar_urf(pattern = "")

pesquisar_cgce3(pattern = "")

pesquisar_cgce2(pattern = "")

pesquisar_cgce1(pattern = "")

pesquisar_cuci_item(pattern = "")

pesquisar_cuci_sub(pattern = "")

pesquisar_cuci_pos(pattern = "")

pesquisar_cuci_cap(pattern = "")

pesquisar_cuci_sec(pattern = "")

pesquisar_isic_classe(pattern = "")

pesquisar_isic_grupo(pattern = "")

pesquisar_isic_div(pattern = "")

pesquisar_isic_sec(pattern = "")
}
\arguments{
\item{pattern}{Trecho de texto ou código para buscar (opcional).}
}
\value{
Data frame com colunas específicas de cada tabela ou mensagem de erro.
}
\description{
Pesquisa padrões de códigos e descrições das tabelas auxiliares do Comex Stat.
Use os wrappers `pesquisar_ncm()`, `pesquisar_sh6()` etc. para cada tabela.
}
