% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consulta_API.R
\encoding{UTF8}
\name{pesquisar_comex_stat}
\alias{pesquisar_comex_stat}
\title{Pesquisa informações no sistema Comex Stat}
\usage{
pesquisar_comex_stat(ano_inicial = "ano corrente" , ano_final = "ano corrente",
mes_inicial = "01", mes_final = "12", detalha_mes = FALSE,  tipo_op ="exp",
tipo_ord= "val",filtros = " ", filtros_esp = " ", detalhamentos = " ", faixa = FALSE,
valor_FOB = TRUE, valor_kg = TRUE, qtd_est = FALSE)
}
\arguments{
\item{ano_inicial}{[int] ano inicial da consulta. ex:2018}

\item{ano_final}{[int] ano final da consulta. ex:2018}

\item{mes_inicial}{[int] mês inicial da consulta. ex:1}

\item{mes_final}{[int] mês inicial da consulta. ex:12}

\item{detalha_mes}{[logical] detalhar informacoes por mês. ex: TRUE}

\item{tipo_op}{[char] tipo da operação, exportação ou importação. Valores possíveis: 'exp','imp'}

\item{tipo_ord}{[char] tipo de ordenamento, valores ou detalhamento. Valores possíveis: 'val', 'det'}

\item{filtros}{[vetor de char] vetor com os filtros desejados. Valores possíveis: 'pais','blocos','uf', 'via', 'urf',
'ncm', 'sh6', 'sh4', 'sh2', 'secao','cgce3', 'cgce2', 'cgce1', 'cuciit', 'cucisub', 'cucigru', 'cucidiv', 'cucisec'}

\item{filtros_esp}{[vetor de char] especificacao dos filtros escolhidos. Um vetor deve ser informado
para cada filtro com todos os itens a serem filtrados. Lista deve estar na ordem que os filtros foram passados.
Para especificação dos filtros é necessario consultar as tabelas auxiliares \code{\link{Tabelas_Auxiliares}}}

\item{detalhamentos}{[vetor de char] vetor com os detalhamentos desejados. Valores possíveis: 'pais','blocos','uf', 'via', 'urf',
'ncm', 'sh6', 'sh4', 'sh2', 'secao','cgce3', 'cgce2', 'cgce1', 'cuciit', 'cucisub', 'cucigru', 'cucidiv', 'cucisec'}

\item{faixa}{[logical] escolha se NCM será uma cesta ou uma faixa. ex: FALSE}

\item{valor_FOB}{[logical] define se os valores FOB serão mostrados}

\item{valor_kg}{define se os valores de quilogramas líquidos serão mostrados}

\item{qtd_est}{define se a quantidade estatística do item sera mostrada. NCM deve estar entre os detalhamentos para que funcione}
}
\description{
Pesquisa estatísticas de importação e de exportação disponibilizadas pelo sistema Comex Stat e gera como resultado um dataframe.
}
\examples{
pesquisar_comex_stat(ano_inicial = 2018, ano_final = 2018, mes_inicial = 1, mes_final = 12,
                     tipo_op = 'exp', tipo_ord = 'val', filtros = c('pais'), detalhamentos = c('pais','ncm'),
                     filtros_esp = list(c(160,249)))

pesquisar_comex_stat(ano_inicial = 2018, ano_final = 2018, mes_inicial = 1, mes_final = 12,
                    detalha_mes = FALSE, tipo_op = 'exp', tipo_ord = 'val', detalhamentos = c('pais','ncm'),
                    filtros = c(), filtros_esp = c())


}
