% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs_auxiliares.R
\encoding{UTF8}
\name{Tabelas_Auxiliares}
\alias{Tabelas_Auxiliares}
\alias{pesquisar_ncm}
\alias{pesquisar_sh6}
\alias{pesquisar_sh4}
\alias{pesquisar_sh2}
\alias{pesquisar_ncm_sec}
\alias{pesquisar_pais}
\alias{pesquisar_blocos}
\alias{pesquisar_uf}
\alias{pesquisar_mun}
\alias{pesquisar_via}
\alias{pesquisar_urf}
\alias{pesquisar_cgce3}
\alias{pesquisar_cgce2}
\alias{pesquisar_cgce1}
\alias{pesquisar_cuci_item}
\alias{pesquisar_cuci_sub}
\alias{pesquisar_cuci_pos}
\alias{pesquisar_cuci_cap}
\alias{pesquisar_cuci_sec}
\title{Pesquisar tabelas auxiliares}
\usage{
pesquisar_ncm('pattern')

pesquisar_sh6('pattern')

pesquisar_sh4('pattern')

pesquisar_sh2('pattern')

pesquisar_ncm_sec('pattern')

pesquisar_pais('pattern')

pesquisar_blocos('pattern')

pesquisar_uf('pattern')

pesquisar_mun('pattern')

pesquisar_via('pattern')

pesquisar_urf('pattern')

pesquisar_cgce3('pattern')

pesquisar_cgce2('pattern')

pesquisar_cgce1('pattern')

pesquisar_cuci_item('pattern')

pesquisar_cuci_sub('pattern')

pesquisar_cuci_pos('pattern')

pesquisar_cuci_cap('pattern')

pesquisar_cuci_sec('pattern')
}
\arguments{
\item{pattern}{Trecho a ser pesquisado (código ou texto)}
}
\description{
As funções \code{pesquisar_ncm}, \code{pesquisar_sh6}, \code{pesquisar_sh4},
\code{pesquisar_sh2} e \code{pesquisar_ncm_sec} se destinam a pesquisar os códigos referentes ao
Sistema Harmonizado.

As funções \code{pesquisar_cgce3}, \code{pesquisar_cgce2} e \code{pesquisar_cgce1} servem para
pesquisar a Classificação por Grandes Categorias Econômicas (CGCE).

As funções \code{pesquisar_cuci_item}, \code{pesquisar_cuci_sub}, \code{pesquisar_cuci_pos}, \code{pesquisar_cuci_cap}
e pesquisar_cuci_sec  são utilizadas para pesquisar a Classificação Uniforme
para o Comércio Internacional (CUCI).

As funções \code{pesquisar_pais} e \code{pesquisar_blocos} se destinam para a pesquisa de países e blocos
econômicos.

As funções \code{pesquisar_uf}, \code{pesquisar_mun} e \code{pesquisar_urf} servem, respectivamente,
para pesquisar a Unidade da Federação (UF), o município e a Unidade da Receita Federal (URF).

A função \code{pesquisar_via} se destina a pesquisar o meio de transporte.
}
\examples{
 pesquisar_ncm('010')

 pesquisar_sh6('aço')

 pesquisar_sh4('frutas')

 pesquisar_sh2('08')

 pesquisar_ncm_sec('animais')

 pesquisar_pais('ale')

 pesquisar_blocos('mercosul')

 pesquisar_uf('bahia')

 pesquisar_mun('são paulo')

 pesquisar_via('aerea')

 pesquisar_urf('brasilia')

 pesquisar_cgce3('alimentos')

 pesquisar_cgce2('insumos')

 pesquisar_cgce1('consumo')

 pesquisar_cuci_item('ovinos')

 pesquisar_cuci_sub('carne')

 pesquisar_cuci_pos('queijo')

 pesquisar_cuci_cap('café')

 pesquisar_cuci_sec('obras')

}
